refactor: organize classes to own files in /src
